requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
weasyprint==59.0
pandas==2.1.1
python-dateutil==2.8.2
Werkzeug==2.3.7